+++ modules/hanna-react/src/Picture.tsx
@@ -1,15 +1,22 @@
 import React from 'react';
 
-import { Image, ImagePropsLinked } from './_abstract/_Image.js';
+import { Image, ImageProps } from './_abstract/_Image.js';
 import { WrapperElmProps } from './utils.js';
 
-export type PictureProps = ImagePropsLinked & {
+export type PictureProps = ImageProps & {
   contain?: boolean;
   className?: string;
 } & WrapperElmProps;
 
 export const Picture = (props: PictureProps) => (
-  <Image {...props} bem="Picture" modifier={props.contain && 'contain'} />
+  <Image
+    {...props}
+    bem="Picture"
+    modifier={[
+      props.contain && 'contain',
+      props.inline && /\.svg(?:\?|$)/i.test(props.src) && 'inlined',
+    ]}
+  />
 );
 
 export default Picture;